python-backend/requirements.txt
# Web框架
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5

# 数据库
PyMySQL==1.1.0
cryptography==41.0.4

# 爬虫相关
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.11.2
lxml==4.9.3

# 数据处理
pandas==2.0.3
numpy==1.24.3
scipy==1.11.1
scikit-learn==1.3.0
psutil==5.9.5

# 任务调度
APScheduler==3.10.4

# 配置管理
python-dotenv==1.0.0

# 日志
loguru==0.7.0

# HTTP客户端
httpx==0.24.1

# 数据验证
marshmallow==3.20.1
email-validator==2.0.0
validators==0.20.0

# 缓存
redis==4.6.0

# 图像处理（用于验证码识别）
Pillow==10.0.0

# 其他工具
fake-useragent==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1